# Some improvements to OME-TIFF write performance