@NicoKiaru 2024-09-24
This [fixes #4204](https://github.com/ome/bioformats/issues/4204#issuecomment-2370738788)!