@ebremer 2024-09-30
@melissalinkert looking forward to testing this!  Now, can you coax some more speed out of the readers? :-)