@sbesson 2024-10-07
As discussed earlier at the weekly @ome/formats meeting, the testing above indicates this PR improves the writing behavior for the two original issues. In preparation for inclusion in the 8.0.0 release, my plan is to run a few additional checks in the upcoming days and cover other typical TIFF writing scenarios including:
- writing default TIFF files
- tiled TIFF writing
- tile sizes & compression
- writing single OME-TIFF vs multiple OME-TIFF vs OME-TIFF with companion

and confirm these changes create no regression.