@sbesson 2024-10-08
Tested the WSI scenario using a synthetic fake file "slide&sizeX=40000&sizeY=40000&sizeC=3&rgb=3&resolutions=5.fake". The data was converted using tiled TIFF with JPEG compression using:
- plain TIFF
- OME-TIFF with 1024x1024 tiles
- OME-TIFF with 512x512 tiles

| Bio-Formats    	|      TIFF      	| OME-TIFF (1024x104) 	| OME-TIFF (512x512) 	|
|----------------	|:--------------:	|---------------------	|--------------------	|
| 7.3.1          	| 181.0 +/- 1.5s 	| 184.5 +/ 4.0s       	| 301.2s +/ 62.3s    	|
| 8.0.0-SNAPSHOT 	| 187.1 +/- 8.9s 	| 177.0s +/- 7.7s     	| 297.1s +/- 47.0s   	|